refactor(desktop): 用 useComposerController 与 useMessageRewind 收拢输入区逻辑
Phase 3：提取 composer/rewind hooks 与 BranchCheckoutDialog 组件。
